requirements.txt
appdirs==1.4.3
Django==1.10.6
django-filter==1.0.1
djangorestframework==3.6.2
Markdown==2.6.8
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7
pyparsing==2.2.0
six==1.10.0
validate-email==1.3
yolk==0.4.3
yolk3k==0.9
